Change AdFailedToLoad mediatorErrorCode type from Long to Int
This improves the developer experience by accepting Int directly from ad networks like AdMob, which provide error codes as Int, rather than requiring developers to manually convert to Long.

Changes:
- AdFailedToLoadData.mediatorErrorCode: Long? -> Int?
- AdEvent.FailedToLoad.mediatorErrorCode: Long? -> Int?
- BackendEvent.Ad.mediatorErrorCode: Long? -> Int?
- Updated all related tests to use Int values

+++ purchases/src/main/kotlin/com/revenuecat/purchases/common/events/BackendEvent.kt
@@ -123,7 +123,7 @@ internal sealed class BackendEvent : Event {
 
         // Failed to load event only fields
         @SerialName("mediator_error_code")
-        val mediatorErrorCode: Long? = null,
+        val mediatorErrorCode: Int? = null,
     ) : BackendEvent()
 
     /**
